+++ vite-project/src/App.jsx
@@ -1,11 +1,5 @@
 import { ShoppingList } from './components/shoppingList/shoppingList'
 
-function MyButton() {
-    return (
-        <button>Im a button</button>
-    )
-}
-
 const user = {
     name: "Alex",
 };
@@ -19,8 +13,7 @@ export default function MyApp() {
             whiteSpace: "pre-wrap"
             }}>list of {user.name} products
           </pre>
-
-          <MyButton /> 
+          <button>Im a button</button>
           <ShoppingList />
         </>
