+++ vite-project/src/mocks/productsMocks.js
@@ -1,0 +1,5 @@
+export const products = [
+    { title: 'Cabbage', isFruit: false, id: 1 },
+    { title: 'Garlic', isFruit: false, id: 2 },
+    { title: 'Apple', isFruit: true, id: 3 },
+  ];
